+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Query params `_q` and `_c` so it's possible to Google Analytics to track the searched terms and category
 
 ## [3.63.1] - 2019-08-21
 ### Fixed